extract manifestDataManager into new script
